Support TAG_Int_Array, support empty lists of TAG_End
Parsing a recent save file from modded Minecraft, I encountered
two difficulties. The first was the lack of support for the new
tag type 11, "TAG_Int_Array" according to a wiki. The second
was that there were a couple of instances of zero-length lists
of type TAG_End. I have no idea whether the fix for that is
correct, but I can now read and write that save file and get
identical output.